* using valarray<size_t>
git-svn-id: https://qucs.svn.sourceforge.net/svnroot/qucs/trunk@1721 b5b04e8c-4942-46c9-ab4f-83783d557d1c

+++ qucs/qucs-filter-v2/qf_matrix.cpp
@@ -39,7 +39,7 @@ Rarray	operator * (Rarray& V, qf_matrix& M) {
 
 gslice	qf_matrix::submx (unsigned r, unsigned k, unsigned c, unsigned l) {
 
-  valarray <unsigned>  len (2), stride (2);
+  valarray <size_t>  len (2), stride (2);
 
   r --; c--;
   k -= r;
